Handle shutdown events properly. Save device_ended immediately.
The SystemEvents.PowerModeChanged was not invoked when the user logged
out and/or shut down the computer. Added SessionEnding subscription to
SystemEventsService.

Change the handling of deviceUsage ended, so that they are saved locally to the
database immediately. Previously, we tried to send it to the online
server, and online on errors would it save the usage locally. There was
never time enough for that to happen.

The device_ended will be send by the Resender at a later time.